requirements.txt
streamlit
vosk
sounddevice
requests
python-dotenv
pyttsx3
